move early rewriting to the semantics package
